IGNITE-21018 MoReview note - move catalogServicice() to TestCluster.
