+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/framework/TestNode.java
@@ -28,7 +28,6 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import org.apache.calcite.tools.Frameworks;
-import org.apache.ignite.internal.catalog.CatalogService;
 import org.apache.ignite.internal.manager.IgniteComponent;
 import org.apache.ignite.internal.sql.engine.InternalSqlRow;
 import org.apache.ignite.internal.sql.engine.QueryCancel;
@@ -76,7 +75,6 @@ public class TestNode implements LifecycleAware {
     private final PrepareService prepareService;
     private final ExecutionService executionService;
     private final ParserService parserService;
-    private final CatalogService catalogService;
 
     private final List<LifecycleAware> services = new ArrayList<>();
 
@@ -96,14 +94,12 @@ public class TestNode implements LifecycleAware {
             MappingService mappingService,
             ExecutableTableRegistry tableRegistry,
             DdlCommandHandler ddlCommandHandler,
-            SystemViewManager systemViewManager,
-            CatalogService catalogService
+            SystemViewManager systemViewManager
     ) {
         this.nodeName = nodeName;
         this.parserService = parserService;
         this.prepareService = prepareService;
         this.schemaManager = schemaManager;
-        this.catalogService = catalogService;
 
         TopologyService topologyService = clusterService.topologyService();
         MessagingService messagingService = clusterService.messagingService();
@@ -227,10 +223,6 @@ public class TestNode implements LifecycleAware {
 
             await(cursor.requestNextAsync(1));
         }
-    }
-
-    public CatalogService catalogService() {
-        return catalogService;
     }
 
     private BaseQueryContext createContext() {
